cmd/meross-device: add configure command
As part of this I used json.Decoder to enable the support for a test to
have multiple commands be returned. This might need some more thought
but is good enough for the scenario where we do a GET on System.All to
determine the MAC address of the device to calculate the password.