cmd/meross-device: add raw command
To help with random testing I decided to add a `raw` command to help
with running custom payloads against the device.

In my manual testing I found that an malformed or bad request will
likely cause the device to reboot.

I couldn't think of a quick way to test this so my manual testing was
all it got for now. I may consider removing this in the future if
additional commands fill the gaps.