cmd/meross-cloud: add cloud tool
This commit adds a simple tool to get existing credentials from the
Meross cloud.

Other tools exist for this but I thought it was worth adding for
completness. The main purpose of this tool is to get the `userID` and
`key` of existing devices that have been provisioned via iot.meross.com.