cmd/meross-device: fix missing -mqtt-host flag err
After trying to run the command I got a bit confused by the flag `-host`
not being correct. Turns out the message didn't match the flag name.

Should probably look at using a CLI package to sort this out but for now
not needing external deps is kind of nice.