home-assistant: add home assistant component
Just added power support and have been using this in Home Assistant
successfully for over a year. Figured I might as well add it to this
repo.

The code isn't clean and probably could be improved a lot to add support
for other Meross devices but it works for me.

The next thing to do would probably be to create a quick start guide for
those that want to hit the ground running but we'll see if there is any
interest first.